Adjust versioning in Gemfile for reliability
I've removed the `shopify_app` version and we can always expect this to
be the lastest. But I've also added a ruby version, based on the minum
required for the gem (which seems sensible).

Quite a bit to remember for future changes, but I think as there is an
example app, keeping it in working order is a good idea!