Failed attempt to use a dedicated action to set cookie
I thought it would be nice to have an action that all it did was induce
user interaction and set the cookie, then redirected back. This would
avoid some of the weirdness around the redirect flow at the moment.

I think it would work in principle, but it gets a little hacky because
the most reliable way to direct to embedded is via EASDK, but the init
method doesn't allow you to redirect to a particular page, so it got
complicated.