Renames generic session to auth_session
We assign many different contexts to the `session` variable
throughout the code-base. This makes it hard to reason about
what context the `session` refers to. In the case of this
change, it's the Shopify Oauth session, but in other areas
it refers to the web session, and the data from the Shop/User
tables. This change begins to make things more explicit.