Improve OAuth tunnel docs
This commit removes duplicate sections on how to set up a tunnel to
allow OAuth redirect during development, makes the one remaining section
clearer and fixes the link it contains to a shopify.dev tutorial.

This gem's docs previously gave instructions to setup a Cloudflared
tunnel, but then directed users to follow a section of a shopify.dev
tutorial which started with setting up an Ngrok tunnel instead before
moving on to adding the redirect URI to the partners dashboard.

After this change, the docs link first to the shopify.dev tutorial
section and include a note after for using Cloudflared should the user
wish to.